+++ contracts/token/psp721/src/extensions/mod.rs
@@ -1,2 +1,3 @@
 pub mod burnable;
 pub mod mintable;
+pub mod metadata;
