+++ contracts/token/psp721/src/traits.rs
@@ -24,10 +24,7 @@ use ink_storage::{
     },
     traits::SpreadLayout,
 };
-pub use psp721_derive::{
-    PSP721MetadataStorage,
-    PSP721Storage,
-};
+
 
 #[cfg(feature = "std")]
 use ink_storage::traits::StorageLayout;
@@ -44,15 +41,6 @@ pub struct PSP721Data {
 }
 
 declare_storage_trait!(PSP721Storage, PSP721Data);
-
-#[derive(Default, Debug, SpreadLayout)]
-#[cfg_attr(feature = "std", derive(StorageLayout))]
-pub struct PSP721MetadataData {
-    pub name: Option<String>,
-    pub symbol: Option<String>,
-}
-
-declare_storage_trait!(PSP721MetadataStorage, PSP721MetadataData);
 
 /// The PSP721 error type. Contract will throw one of this errors.
 #[derive(strum_macros::AsRefStr)]
@@ -307,26 +295,6 @@ pub trait PSP721: PSP721Storage {
     fn _burn(&mut self, id: Id) {
         let caller = Self::env().caller();
         self._burn_from(caller, id);
-    }
-}
-
-#[brush::trait_definition]
-pub trait PSP721Metadata: PSP721MetadataStorage {
-    /// Returns the token name.
-    #[ink(message)]
-    fn name(&self) -> Option<String> {
-        self.get().name.clone()
-    }
-
-    /// Returns the token symbol.
-    #[ink(message)]
-    fn symbol(&self) -> Option<String> {
-        self.get().symbol.clone()
-    }
-
-    fn _init_with_metadata(&mut self, name: Option<String>, symbol: Option<String>) {
-        self.get_mut().name = name;
-        self.get_mut().symbol = symbol;
     }
 }
