+++ contracts/token/psp721/tests/tests.rs
@@ -15,6 +15,7 @@ mod tests {
     use psp721::{
         extensions::{
             burnable::*,
+            metadata::*,
             mintable::*,
         },
         traits::*,
